requirements.txt
Flask>=2.0
Pillow
numpy
scikit-learn
scikit-image
joblib
